requirements.txt
# Cotton Leaf Disease Detection System - Root Requirements
# This file lists all dependencies for quick setup

# For detailed requirements, see:
# - backend/requirements.txt
# - pc_app/requirements.txt
# - frontend/package.json

# Core ML and CV
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.227
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# Backend API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Data processing
pyyaml==6.0.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0

# Optional: ONNX Runtime
# onnxruntime-gpu==1.16.3
# onnx==1.15.0

# Optional: Data augmentation
# albumentations==1.3.1

# Optional: Visualization
# matplotlib==3.8.2
# seaborn==0.13.0

# Development
# pytest==7.4.3
# black==23.12.0
# flake8==6.1.0
